test: simplify test_get_blueprints_basic_functionality() a bit
This is just a quick drive-by from having a quick peek at the
mcp service. It tweaks test_get_blueprints_basic_functionality()
a tiny bit so that the test just compares the full response to
the expected string. This covers what the test is currently
doing but feels easier to read (to me).

Alternatively to "spell" the json out as a string we could
of course have a "expected_result" list as python objects
and then json.dump() it but this seems fine.

It also tweaks the mock_api_response() so that the data is
not already sorted so that we can test that the get_blueprints()
actually sorts the reply.